Move constraint collapse tests from separate file to trait resolution test suite
